Kocluk/requirements.txt
bcrypt==4.0.1
cryptography==43.0.0
fastapi==0.112.2
python-jose==3.3.0
python-multipart==0.0.9
SQLAlchemy==2.0.32
uvicorn==0.30.6
passlib==1.7.4
pytest==8.3.2
httpx==0.27.2
pytest-asyncio==0.24.0
aiofiles==24.1.0
jinja2==3.1.4
python-dotenv==1.0.1
google-generativeai==0.7.2
langchain-google-genai==1.0.10
markdown==3.7
bs4==0.0.2
alembic==1.13.2
starlette~=0.38.6
pydantic~=2.11.4
routers~=0.10.1
google~=3.0.0
protobuf~=4.25.7
pyasn1~=0.6.1